Version06
Fixed various errors such as user categories not loading, packages not loading. created Headquarters backend crud, created headquaters frontend crud, with its context, api and used it on RegisterAgent page, also made RegisterAgent page functional, users can now be added to the DB, added a wallpaper responsive wallpaper for every page, Modified users.controllers, before: when you registered and agent it would automatically set the cookies on the new agent, now it doesnt as this was not intended. also made navbar's logo clickeable and it will take you to loginClient page if signed in, if not it will take you to the AgentLogin page